+++ Repay_Bank/Repay_Bank.DataAccessLayer/Migrations/20230825075730_add_confirmCOde.cs
@@ -1,0 +1,26 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Repay_Bank.DataAccessLayer.Migrations
+{
+    public partial class add_confirmCOde : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "ConfirmCode",
+                table: "AspNetUsers",
+                type: "int",
+                nullable: true,
+                defaultValue: 0);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "ConfirmCode",
+                table: "AspNetUsers");
+        }
+    }
+}