+++ Repay_Bank/Repay_Bank.EntityLayer/Concrete/AppUser.cs
@@ -14,6 +14,8 @@ namespace Repay_Bank.EntityLayer.Concrete
         public string District { get; set; }
         public string City { get; set; }
         public string CustomerImageUrl { get; set; }
+        public int ConfirmCode { get; set; }
         public List<CustomerAccount> CustomerAccounts { get; set; }
+        
     }
 }