+++ Repay_Bank/Repay_Bank.PresentationLayer/Program.cs
@@ -1,7 +1,28 @@
+using Repay_Bank.DataAccessLayer.Concrete;
+using Repay_Bank.DTO.DTOS.AppUserDtos;
+using Repay_Bank.EntityLayer.Concrete;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.AddControllersWithViews();
+//
+builder.Services.AddDbContext<Context>();
+builder.Services.AddIdentity<AppUser,AppRole>(
+    opt =>
+    {
+        opt.Password.RequireNonAlphanumeric= false;
+        opt.Password.RequiredLength=3;
+        opt.Password.RequireUppercase = false;
+        opt.Password.RequireLowercase = false;
+		opt.Password.RequireDigit = false;
+		// opt.SignIn.RequireConfirmedEmail = true;//mail doðrulamasý olsun mu
+
+	}
+	)
+    .AddErrorDescriber<CustomerIdentityValidation>()    
+    .AddEntityFrameworkStores<Context>();
+
 
 var app = builder.Build();
 
@@ -18,6 +39,7 @@ app.UseStaticFiles();
 
 app.UseRouting();
 
+app.UseAuthentication();
 app.UseAuthorization();
 
 app.MapControllerRoute(